+++ packages/core/src/GetCandyServiceProvider.php
@@ -171,6 +171,10 @@ class GetCandyServiceProvider extends ServiceProvider
                 ], 'getcandy');
             });
 
+            $this->publishes([
+                __DIR__.'/../database/migrations/' => database_path('migrations')
+            ], 'getcandy-migrations');
+
             $this->commands([
                 InstallGetCandy::class,
                 AddonsDiscover::class,