+++ packages/admin/CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### [Unreleased]
+
+### Added
+
+- Migrations are now publishable under the tag `getcandy-migrations`
+
 ### 2.0-beta13.2 - 2022-06-23
 
 ## Fixed